Add extremely crude views.
For some reason, the ruby plugin API is refusing to allow me to use stapler
views for the build action; it is forcing the action into rack integration
mode. I'm not sure how to get the rest of Jenkins' chrome up in this mode.

This implementation will suffice for a demo, though.